Reset values in migration
